services/api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
pypdf
langchain
langchain-community
langchain-google-genai
chromadb
